Fixed latency issue

+++ Servo.hpp
@@ -124,7 +124,7 @@ class Servo {
 
             // printf("Angle requested: %f deg, after adding offset: %f deg\n", angle, rel_angle_deg);
 
-            rel_angle_deg = fmodf(rel_angle_deg, 360.0f);
+            rel_angle_deg = fmod(rel_angle_deg, 360.0f);
 
             // printf("Angle between -360 to 360: %f deg\n", rel_angle_deg);
